### Draft Commit Message
```
refactor(menu): implement navigation stack and restructure menu items

- Add menu navigation stack (push/pop) to support submenu transitions
  with state restoration (scroll offset, selected index)
- Restructure MenuItem_t: move value_limits outside union for consistent
  access, add on_value_changed callback for status bar updates
- Add GUI_InvalidateStatusBar() for forced status bar redraw
- Move global settings variables to gui.c (bluetooth, WiFi, CC1101 params)
- Expand DS3231 square wave frequency parameter from uint8_t to uint16_t
- Clean up: remove mock measurement module, delete OLD.zip, remove
  dead/commented-out code (font metrics, UI layout)
- Increase clock width in status bar from 60 to 70
```

### Conventional Commit (short)

```
refactor(menu): implement navigation stack and restructure menu items
```

Хотите, чтобы я закоммитил эти изменения с этим сообщением?

+++ lib/DS3231/ds3231.c
@@ -191,7 +191,7 @@ float DS3231_GetTemperature(I2C_HandleTypeDef *hi2c) {
 }
 
 // Включение квадратного генератора (частота: 1, 1024, 4096, 8192 Гц)
-DS3231_Status_t DS3231_EnableSquareWave(I2C_HandleTypeDef *hi2c, uint8_t frequency) {
+DS3231_Status_t DS3231_EnableSquareWave(I2C_HandleTypeDef *hi2c, uint16_t frequency) {
     uint8_t control;
     DS3231_Status_t ret = DS3231_ReadReg(hi2c, DS3231_REG_CONTROL, &control);
     if (ret != DS3231_OK) return ret;
